Update patch set 1
Patch Set 1: Fails; No score

Independently from technical considerations, we cannot accept this change in its exact current state, because the only explicit copyright notice that we allow in common Android files is "The Android Open-Source Project".

Thank you for uploading a replacement change so that we can proceed with the rest of the review process. Because the change involves copyright notices, we can't make that change on your behalf.

Patch-set: 1