Crash on XHTML pages with form element but without namespace URI.
When an XHTML page lacks the 'xmlns' attribute in its 'html' element,
it results in the parser constructing a DOM tree where each
node is of a generic Element type. This means that it is not
correct to extract nodes from that tree and then cast the pointers
to HTMLELement types. To avoid the crash, I have introduced
checks that the nodes in question are indeed HTML elements.

Change-Id: If3ae4e517e7bf00948018cbbc2ffbf3d0facb1aa