Update patch set 2
Patch Set 2: No score; Looks good to me, approved

The original code was a misunderstanding of the interface, and assumed that the point for drawing (the baseline) was the same as the point provided for computing the bounds which is already set to the upper left.

This change is a good approximation, but the code is incomplete and only works for simple cases. A better code model is provided in the Chromium port here:
http://trac.webkit.org/browser/trunk/WebCore/platform/graphics/chromium/FontLinux.cpp

Patch-set: 2
Reviewer: Cary Clark <1002451@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=0