Bug fix 2337042 <base> URL + <img> URL = URL that starts with "/."
The Android browser inserts "/." at the beginning of the URL path
when processing an <img>  in an HTML document with a <base> tag.
This does not cause any errors but should be fixed to comply with
RFC 3986 section 5.4.

This only happens at the first pass of the URL parsing, at the second
pass where the base and path are parsed together this is dealt with
correctly.  I am therefore fixing it by modifying the parse() method
in KURL to check for a trailing '.' at the end of the base segment.

Change-Id: I0d641c8619e00aa2b08a33e271ef863cee5ed7ad