Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Need to review and discuss this fix. When multiple browser windows are closed and opened, once in a while, the garbage collector frees up the memory of the java webview core object and the native webview core finds null reference for the JNI method. There needs to be some synchronization between JAVA and native. This fix ensure JAVA object is not cleaned up and crash is avoided. Is this the right fix or could we do something better?

Patch-set: 1