Allow scrolling of framesets even when frameflattening is enabled
On a page that uses frames, and with a touch screen, it is
difficult to distinguish between scrolling an individual frame
and scrolling the entire page. To get rid of this
problem, Android uses something called "frame flattening".
This means that a frame is expanded to the size of its content
and hence there is nothing to scroll in the frame any more.

Normally, on a page with frames, Webkit disables scrolling for
the main frameset. However, such scrolling should not be disabled
if frame flattening is in effect, since then the user ends up not
being able to scroll at all: neither for the main frameset nor
for the individual frames.

When disabling scrolling for the main frameset, a check is first
made to see if frame flattening is active. However, Android's
frame flattening code does not use the function
frameSetFlatteningEnabled(), instead it is governed by a define,
FLATTEN_FRAMESET. As a result, scrolling was incorrectly disabled
for the main frameset even though Android's frame flattening was
used.

Change-Id: I49b27c263e4a2e837a2285ce2b243dc4749105ab