Enable html selects to be rendered correctly even at small sizes.
The 9-patch used for comboboxes previously didn't allow the
box to be rendered with a smaller height than the pixelheight of
the graphical resource. Together with an error in the code, that
should limit the height, this caused small comboboxes to be
rendered incorrect and impossible to press.

To solve the problem, the arrow of the combobox has been broken
out of the combobox graphical resource and put into an image of
its own. The 9-patch area has been expanded to allow the
graphical resource to scale down in height. The combobox arrow
is always placed in the vertical center and drawn after the
scaling has occurred. The extra padding on the top and bottom of
the combobox has been removed, as this caused layout errors,
e.g. errors with non aligned combobox text fields, when the
combobox size was smaller than the graphical resource.

Change-Id: I83947f3da38c70e889f4b0a5b6389300da081ba1