Update patch set 1
Patch Set 1:

Note that the v8 change has issues, and that this change depends on it.

My best guess is that this change doesn't cause any difficulties (other than the dependency on v8). It is fine in the AOSP master, and compiles fine in Google's internal master (doesn't link for lack of v8).

Patch-set: 1