Fix race condition issue in TilesManager
The code of creating TilesManager instance is multi-thread unsafe. It's a typical singleton issue.
When constructing the TilesManager instance, a new TexturesGenerator thread will run into the same code, that
cause the race condition in browser. The browser is hanging on starting stage and pop out "not responsing" on
x86 emulator image. This patch uses double-checked multi-thread safe code to fix the issue.