Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

To answer steve's question, the way we may use the code, it is definitely a possibility. Afaik, on master we did remove calls that ended up creating the singleton at the wrong time, but we should still protect the creation of the singleton (probably using Steve's suggestion than this CL's approach).

Patch-set: 1
Label: Code-Review=-1