Update patch set 2
Patch Set 2:

we have measured the browser fps by adding some debug code to surfacetexture. and found:
when in fullscreen mode, the browser should have no activity,
but in fact, the timer event will trigger the controld bar update then will cause browser do composition and send updated surface to surfaceflinger.
For example: play a 30 fps video will produce ~30fps fullscreen surfaceview update which is visiable and ~10fps browser update which is invisiable.
when in inline mode, this will produce ~40fps for browser.

and this patch only reduce the display update, the new position will be logged, when visible. it will use the new position and display it.

Patch-set: 2
Label: Code-Review=0