Update patch set 2
Patch Set 2:

Thanks for the details. This is a good improvement. But if this is targeted to help the full screen only, then it could be limited to that. Maybe you can limit the time update message when we are in full screen mode from Java side?

BTW, I locally applied this CL, and I do see the jump myself in inline mode.
Basically, if you go to nytimes.com which is normally a long HTML5 video (>1 min). Now start inline mode and let it continue for > 30 seconds. Before it end, touch the video, the progress bar will show up, then you will see the thumb is jumping from left(before the hiding) to the right( the current position). It is just a visual glitch.
For inline mode, if we have to do a trade off here, I think we had better avoid this jump.

Patch-set: 2