Update patch set 1
Patch Set 1:

Nobody can call AddRef if the ref count is 0 unless someone is using raw pointers or some unsynchronized global or singleton which would be a different bug.

But what I think happens is this:
1: Thread A creates the FlushSemaphore (refcnt = 1)
2: A calls SendFlushRequest
3: SendFlushRequest incs refcnt, sends pointer to thread B (refcnt = 2?)
4: Thread B does the work, decs the refnt (refcnt = 0?)

Due to the lack of using locks/atomics, the inc in step 3 on Thread A didn't get flushed. Thus, the dec ref in step 4 thinks the ref count is now 0 and destroys it although the real count should actually be 1.

Is there something else I'm not seeing about this code?

Patch-set: 1