Update patch set 1
Patch Set 1:

"Due to the lack of using locks/atomics, the inc in step 3 on Thread A didn't get flushed.".
I suppose you mean it needs memory barrier here. Right?

We will try to reproduce it with the change you suggested. But I can't commit when we could make it happen.

Patch-set: 1