Update patch set 1
Patch Set 1: Do not submit

So having looked at this it is not a change we want. What's happening on that site is the inner frame is expanded correctly, however the outer frame is not because the <frame> tag has scrolling="NO", which we are respecting. It's unfortunate that the inner frame is being clipped by the outer frame, but the change to basically ignore scrolling="NO" isn't really the right fix. That's trading one bug for another :)

The right fix would be to disable flattening of frames and instead make them composited scrolling layers like we do for overflow divs and iframes. Then the inner frame would scroll as expected, without breaking scrolling="NO"

I tested nested frames without scrolling="NO" and I am not seeing any scrolling issues in that case. The inner frames correctly make the outer ones bigger, and everything is flattened completely.

Patch-set: 1
Label: Code-Review=-2