Update patch set 1
Patch Set 1:

The patch is needed, because correct libwebcore.a creation relies on non-standard ar’s command-line argument ‘P’.

Intel compiler's wrapper of ar for multi-file optimization tries to be compatible with GNU ar but currently 'P' option is not fully supported.

Per earlier comment on this patch, renamed Android-specific MIMETypeRegistry.cpp to MIMETypeRegistryAndroid.cpp. The change for Connection.cpp is similar.

Names of the files are in line with the names of other platform-specific MIMETypeRegistry.cpp and Connection.cpp

Patch-set: 1